refactor: move installer/ and launcher/ under src/ with colocated tests
Move all TypeScript source into src/installer/ and src/launcher/, colocate
test files beside the source they test (removing test/ subdirectories), and
update all tooling references in build.ts, tsconfig.json, and package.json
scripts. Update relative imports in moved test files, fix the launcher-install
path to src/launcher/myclaude.sh, and add an extra dirname call in
src/installer/main.ts so scriptDir correctly resolves to the repo root.
