+++ src/installer/claude.ts
