+++ src/installer/cli.ts
