+++ src/installer/colors.ts
