+++ src/installer/constants.ts
