+++ src/installer/mcp.ts
