+++ src/launcher/cloudwatch.test.ts
@@ -3,7 +3,7 @@ import assert from "node:assert/strict";
 import * as fs from "node:fs";
 import * as path from "node:path";
 import * as os from "node:os";
-import { loadAwsProfiles } from "../mcp/cloudwatch.js";
+import { loadAwsProfiles } from "./mcp/cloudwatch.js";
 
 // ---------------------------------------------------------------------------
 // Shared temp directory — cleaned up after all tests complete
