+++ tsconfig.json
@@ -9,5 +9,5 @@
     "skipLibCheck": true,
     "types": ["node"]
   },
-  "include": ["installer/**/*.ts", "launcher/**/*.ts", "build.ts"]
+  "include": ["src/installer/**/*.ts", "src/launcher/**/*.ts", "build.ts"]
 }
